Completed base for MapManager!!! It is now able to generate a simple random map (code well structured with proper data structure (hashmap/dictionary) and comments)
