--- Assets/Scripts/MapManagerV2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6ab0bb2a631b39046994d3f39522a596
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
