--- Assets/Scripts/Path.cs
@@ -1,22 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Path : MonoBehaviour
-{
-    // [SerializeField] private Room a;
-    // [SerializeField] private Room b;
-    // private bool oneWay;
-    
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
