--- Assets/Scripts/Path.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 88eb9b9cf5ecbbe4a973a1b8aa15fe41
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
