--- Assets/Scripts/Room.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ad99995abd41544ebcfafe2c625e835
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
